third_party/blink/web_tests/external/wpt/audio-session/audiosession-default-values.https-expected.txt

This is a testharness.js-based test.
[FAIL] navigator.audioSession exposes an AudioSession instance
  assert_implements: navigator.audioSession is not supported undefined
[FAIL] AudioSession.type defaults to 'auto'
  assert_implements: navigator.audioSession is not supported undefined
[FAIL] AudioSession.state defaults to 'inactive'
  assert_implements: navigator.audioSession is not supported undefined
[FAIL] navigator.audioSession is [SameObject]
  assert_implements: navigator.audioSession is not supported undefined
Harness: the test ran to completion.
